fix: ohe datatype

+++ aucmedi/data_processing/io_interfaces/io_csv.py
@@ -124,7 +124,7 @@ def csv_loader(path_data, path_imagedir, allowed_image_formats,
         class_names = np.unique(classes_sparse).tolist()
         class_n = len(class_names)
         # Parse sparse categorical annotations to One-Hot Encoding
-        class_ohe = pd.get_dummies(classes_sparse).to_numpy()
+        class_ohe = pd.get_dummies(classes_sparse, dtype=np.uint8).to_numpy()
     # Try parsing one-hot encoded format (CSV Format 2)
     else:
         # Identify OHE columns
